pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "django-cognito-m2m"
version = "0.1.0"
description = "Django and DRF integration for AWS Cognito machine-to-machine bearer token authentication."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE.txt"}
authors = [
  {name = "Caltech IMSS ADS", email = "imss-ads-staff@caltech.edu"}
]
keywords = ["aws", "cognito", "oauth2", "client credentials", "access token validation"]
dependencies = [
  "Django>=4.2",
  "m2m-cognito",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Framework :: Django",
  "Framework :: Django :: 4.2",
  "Framework :: Django :: 5.0",
  "Framework :: Django :: 5.1",
  "Framework :: Django :: 5.2",
  "Framework :: Django :: 6.0",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Programming Language :: Python :: 3.15",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Security",
]

[project.urls]
Homepage = "https://github.com/caltechads/django-cognito-m2m"

[project.optional-dependencies]
drf = ["djangorestframework>=3.15"]
test = [
  "pytest>=8.0",
  "pytest-django>=4.8",
  "djangorestframework>=3.15",
]

[tool.hatch.build.targets.wheel]
packages = ["django_cognito_m2m"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["tests"]

[tool.bumpversion]
current_version = "0.1.0"
tag = true
tag_name = "{new_version}"
commit = true
message = "Bump version: {current_version} -> {new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "django_cognito_m2m/__init__.py"